Updated swift sample codes
