+++ Example/ObjectiveC Example/Channel Plugin Sample objc.xcworkspace/xcuserdata/haeunchung.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -58,12 +58,12 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "Channel Plugin Sample objc/AppDelegate.m"
-            timestampString = "526382967.365512"
+            timestampString = "544446017.921088"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "74"
-            endingLineNumber = "74"
-            landmarkName = "-application:didReceiveRemoteNotification:fetchCompletionHandler:"
+            startingLineNumber = "69"
+            endingLineNumber = "69"
+            landmarkName = "-userNotificationCenter:willPresentNotification:withCompletionHandler:"
             landmarkType = "7">
          </BreakpointContent>
       </BreakpointProxy>
@@ -74,12 +74,12 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "Channel Plugin Sample objc/AppDelegate.m"
-            timestampString = "526382967.365512"
+            timestampString = "544446017.921192"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "69"
-            endingLineNumber = "69"
-            landmarkName = "-application:didReceiveRemoteNotification:"
+            startingLineNumber = "64"
+            endingLineNumber = "64"
+            landmarkName = "-application:didReceiveRemoteNotification:fetchCompletionHandler:"
             landmarkType = "7">
          </BreakpointContent>
       </BreakpointProxy>
@@ -288,6 +288,22 @@
             startingLineNumber = "348"
             endingLineNumber = "348"
             landmarkName = "isChannelPushNotification(_:)"
+            landmarkType = "7">
+         </BreakpointContent>
+      </BreakpointProxy>
+      <BreakpointProxy
+         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
+         <BreakpointContent
+            shouldBeEnabled = "Yes"
+            ignoreCount = "0"
+            continueAfterRunningActions = "No"
+            filePath = "Channel Plugin Sample objc/UserViewController.m"
+            timestampString = "544446017.9218301"
+            startingColumnNumber = "9223372036854775807"
+            endingColumnNumber = "9223372036854775807"
+            startingLineNumber = "54"
+            endingLineNumber = "54"
+            landmarkName = "-didClickOnBoot:"
             landmarkType = "7">
          </BreakpointContent>
       </BreakpointProxy>
