+++ Example/ObjectiveC Example/Channel Plugin Sample objc/AppDelegate.h
@@ -12,6 +12,5 @@
 
 @property (strong, nonatomic) UIWindow *window;
 
-
 @end
